+++ nodeconductor/backup/tests/factories.py
@@ -45,7 +45,7 @@ class BackupFactory(factory.DjangoModelFactory):
         self.metadata.update(
             {
                 'cloud_project_membership': self.backup_source.cloud_project_membership.pk,
-                'hostname': 'original.vm.hostname',
+                'name': 'original.vm.name',
                 'template': template.pk,
                 'system_snapshot_id': self.backup_source.system_volume_id,
                 'system_snapshot_size': self.backup_source.system_volume_size,